chore: rename lib-franklin
