+++ 404.html
@@ -11,7 +11,7 @@
   <meta property="og:title" content="Page not found">
   <script src="/scripts/scripts.js" type="module" crossorigin="use-credentials"></script>
   <script type="module">
-    import { sampleRUM } from '/scripts/block-utils.js';
+    import { sampleRUM } from '/scripts/lib-franklin.js';
 
     window.addEventListener('load', () => {
       if (document.referrer) {
