+++ blocks/footer/footer.js
@@ -1,4 +1,4 @@
-import { readBlockConfig, decorateIcons } from '../../scripts/block-utils.js';
+import { readBlockConfig, decorateIcons } from '../../scripts/lib-franklin.js';
 
 /**
  * loads and decorates the footer
