+++ scripts/lib-franklin.js
