+++ test/scripts/block-utils.test.js
@@ -12,7 +12,7 @@ document.head.innerHTML = await readFile({ path: './head.html' });
 
 describe('Utils methods', () => {
   before(async () => {
-    blockUtils = await import('../../scripts/block-utils.js');
+    blockUtils = await import('../../scripts/lib-franklin.js');
     document.body.innerHTML = await readFile({ path: './body.html' });
   });
